--- a/src/Toml/Type/TOML.hs
+++ b/src/Toml/Type/TOML.hs
@@ -13,6 +13,7 @@ module Toml.Type.TOML
 import Control.DeepSeq (NFData)
 import Data.HashMap.Strict (HashMap)
 import Data.List.NonEmpty (NonEmpty)
+import Data.Maybe (isJust)
 import Data.Semigroup (Semigroup (..))
 import GHC.Generics (Generic)
 
@@ -20,7 +21,7 @@ import Toml.PrefixTree (Key (..), PrefixMap)
 import Toml.Type.AnyValue (AnyValue (..))
 import Toml.Type.Value (Value)
 
-import qualified Data.HashMap.Strict as HashMap
+import qualified Data.HashMap.Strict as H
 import qualified Toml.PrefixTree as Prefix
 
 
@@ -89,17 +90,29 @@ data TOML = TOML
     { tomlPairs       :: HashMap Key AnyValue
     , tomlTables      :: PrefixMap TOML
     , tomlTableArrays :: HashMap Key (NonEmpty TOML)
+    , failure         :: Maybe String
     } deriving (Show, Eq, NFData, Generic)
 
 instance Semigroup TOML where
-    (TOML pairsA tablesA arraysA) <> (TOML pairsB tablesB arraysB) = TOML
-        (pairsA <> pairsB)
-        (HashMap.unionWith (<>) tablesA tablesB)
-        (arraysA <> arraysB)
+    (TOML pairsA tablesA arraysA fA) <> (TOML pairsB tablesB arraysB fB)
+        | isJust fB = TOML mempty mempty mempty fB
+        | isJust fA = TOML mempty mempty mempty fA
+        | not (null kv) = TOML mempty mempty mempty (Just kvErr)
+        | not (null ta) = TOML mempty mempty mempty (Just taErr)
+        | otherwise = TOML (pairsA <> pairsB)
+                           (H.unionWith (<>) tablesA tablesB)
+                           (arraysA <> arraysB)
+                           Nothing
+      where
+        kv = H.intersection pairsA pairsB
+        kvErr = "Keys defined multiple times: " ++ show (H.keys kv)
+        ta = H.intersection arraysA arraysB
+        taErr = "Keys defined multiple times: " ++ show (H.keys ta)
+
 
 instance Monoid TOML where
     mappend = (<>)
-    mempty = TOML mempty mempty mempty
+    mempty = TOML mempty mempty mempty mempty
 
 -- | Inserts given key-value into the 'TOML'.
 insertKeyVal :: Key -> Value a -> TOML -> TOML
@@ -107,7 +120,7 @@ insertKeyVal k v = insertKeyAnyVal k (AnyValue v)
 
 -- | Inserts given key-value into the 'TOML'.
 insertKeyAnyVal :: Key -> AnyValue -> TOML -> TOML
-insertKeyAnyVal k av toml =toml { tomlPairs = HashMap.insert k av (tomlPairs toml) }
+insertKeyAnyVal k av toml =toml { tomlPairs = H.insert k av (tomlPairs toml) }
 
 -- | Inserts given table into the 'TOML'.
 insertTable :: Key -> TOML -> TOML -> TOML
@@ -118,5 +131,5 @@ insertTable k inToml toml = toml
 -- | Inserts given array of tables into the 'TOML'.
 insertTableArrays :: Key -> NonEmpty TOML -> TOML -> TOML
 insertTableArrays k arr toml = toml
-    { tomlTableArrays = HashMap.insert k arr (tomlTableArrays toml)
+    { tomlTableArrays = H.insert k arr (tomlTableArrays toml)
     }