This is a proof of concept to address #5 and #211 (it is already enough to fix 211 in its current form).

This is quite a big issue so I want to make sure this approach is OK before working too much on it. The basic idea is that I'm adding a `failure :: Maybe String` field to `TOML` and tweaks its `(<>)`
definition so that it detects invalid syntax there. If there is a failure, the parsing will fail and return an error.

What do you think?

It passes all test and here are benchmarks:

master:
```
benchmarked Parse/tomland
time                 354.0 μs   (349.8 μs .. 360.0 μs)
                     0.998 R²   (0.996 R² .. 1.000 R²)
mean                 358.5 μs   (355.7 μs .. 362.4 μs)
std dev              11.86 μs   (8.058 μs .. 17.37 μs)
variance introduced by outliers: 15% (moderately inflated)
```

with validation:
```
benchmarked Parse/tomland
time                 376.4 μs   (369.0 μs .. 386.3 μs)
                     0.996 R²   (0.992 R² .. 0.999 R²)
mean                 377.4 μs   (373.9 μs .. 383.4 μs)
std dev              15.26 μs   (11.03 μs .. 22.16 μs)
variance introduced by outliers: 20% (moderately inflated)
```