@chshersh 2019-08-14
@jiegillet Thanks for your effort on working on this issue! Validation is hard, and it's unfortunate, that it decreases the performance, but correctness is more important. Also, I'm not entirely happy with adding `failure` to the `TOML` data type because `TOML` in my vision should represent TOML AST, not the state of parsing.

I see that TOML parser in `tomland` becomes more complicated. For quite some time I was thinking about how to make it simpler, faster and support validation as well. I have some ideas now. Will try to put them in a separate PR.