@chshersh 2019-09-28
@jiegillet I've implemented an alternative solution in #216 (including some validation). So I'm closing this PR in favour of another one.